requirements.txt
# Core web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ML / Data Science
numpy>=1.26.0
pandas
scikit-learn>=1.4.0
joblib==1.5.1

# LLM & embeddings
openai==1.99.9
sentence-transformers
faiss-cpu
# faiss-cpu==1.7.4
# faiss-gpu == 1.7.2
transformers>=4.42.0
nltk>=3.9.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.2

# Profiling / EDA
ydata-profiling>=4.8.3

# Data handling
pyarrow>=14.0.0

# Background tasks / caching
redis==5.0.7
celery==5.4.0

# Monitoring & telemetry
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0

# Config & utilities
pydantic-settings==2.5.2
pydantic==2.11.2
httpx==0.27.0
python-dotenv==1.0.1

# Optional text/NER utils
spacy==3.7.5
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Testing
pytest==8.2.2
anyio==4.4.0

# Lint/formatting
ruff==0.5.0


Flask == 2.3.3
pytz == 2022.1
rich == 13.7.1